Transition code to Jinja2
